+++ src/components/PopupWithImage.js
@@ -1,22 +1,15 @@
 import Popup from './Popup.js';
 
-import {
-    imgPopUp,
-    imgPopUpTitle,
-} from '../components/data.js';
-
-
-
-export default class PopupWithImage extends Popup {
+export class PopupWithImage extends Popup {
     constructor({ data }, selectorPopup) {
         super(selectorPopup);
         this._titleCard = data.name;
         this._imageCard = data.link;
     }
     openPopUp() {
-        imgPopUp.src = this._imageCard;
-        imgPopUpTitle.textContent = this._titleCard;
-        imgPopUp.alt = this._titleCard;
-        super.openPopUp();
+        this._popup.querySelector('.popup__img-title').textContent = this._titleCard
+        this._popup.querySelector('.popup__img').src = this._imageCard
+        this._popup.querySelector('.popup__img').alt = this._titleCard
+        super.openPopUp()
     }
 }
