--- src/components/constants.js
@@ -1,9 +1,0 @@
-// Константы для валидации форм
-export const validationConfig = {
-    formSelector: '.form',
-    inputSelector: '.form__input',
-    submitButtonSelector: '.form__submit',
-    inactiveButtonClass: 'popup__button-save_inactive',
-    //inputErrorClass: 'popup__item_type_error',
-    errorClass: 'form__input-error_active'
-}
