+++ src/components/data.js
@@ -63,3 +63,13 @@ export const title = popupAddCard.querySelector('#name-input')
 export const photoAvatar = document.querySelector('.profile__photo')
 export const formAvatarChange = popUpChahgeAvatar.querySelector('.form')
 export const urlLinkAvatar = popUpChahgeAvatar.querySelector('#url-input')
+
+// Константы для валидации форм
+export const validationConfig = {
+  formSelector: '.form',
+  inputSelector: '.form__input',
+  submitButtonSelector: '.form__submit',
+  inactiveButtonClass: 'popup__button-save_inactive',
+  //inputErrorClass: 'popup__item_type_error',
+  errorClass: 'form__input-error_active'
+}
